Move the logic to DonateViewModel
